refactor: env-configurable API base URL, move routeCache to services/
- Add apiBaseUrl to config (API_BASE_URL env var, default https://api.lad.lviv.ua)
- apiUrl.js reads from config instead of hardcoded constant
- Move utils/routeCache.js -> services/routeCache.js (it is a stateful service, not a utility)
- Update routes/index.js import and test/routeCache.test.js import accordingly
- Add config test for apiBaseUrl default
