+++ config/index.js
@@ -6,4 +6,5 @@ module.exports = {
     pollIntervalMs: (parsedPollInterval > 0) ? parsedPollInterval : 5000,
     port: parseInt(process.env.PORT) || 3000,
     googleMapsKey: process.env.GOOGLE_MAPS_KEY || '',
+    apiBaseUrl: process.env.API_BASE_URL || 'https://api.lad.lviv.ua',
 };
